Update remote-login.sh
Run this script on a computer, and you will be able to ssh into it if you have a username and password. Note that this is not great for long term security, but it's great for a quick setup where you need ssh right away. Router port forwarding will likely be required.